Script to Import chats, users and messages from echo and sl.

+++ packages/cli/tool/scripts/echo-sl-compare.sh
@@ -3,14 +3,22 @@
 mkdir -p $1
 cd $1
 
-wget 'https://www.sqlite.org/2022/sqlite-tools-osx-x86-3400100.zip'
-unzip sqlite-tools-osx-x86-3400100.zip
-rm sqlite-tools-osx-x86-3400100.zip
-mv sqlite-tools-osx-x86-3400100 sqlite-tools-osx
+DIR=sqlite-tools-osx
+if [ ! -d "$DIR" ]; then
+    wget 'https://www.sqlite.org/2022/sqlite-tools-osx-x86-3400100.zip'
+    unzip -q sqlite-tools-osx-x86-3400100.zip
+    rm sqlite-tools-osx-x86-3400100.zip
+    mv sqlite-tools-osx-x86-3400100 $DIR
+fi
 
-wget 'https://sbp.enterprisedb.com/getfile.jsp?fileid=1258319' -O pgsql.zip
-unzip pgsql.zip
-rm pgsql.zip
+DIR=pgsql
+if [ ! -d "$DIR" ]; then
+    wget 'https://sbp.enterprisedb.com/getfile.jsp?fileid=1258319' -O pgsql.zip
+    unzip -q pgsql.zip
+    rm pgsql.zip
+fi
+
+two_weeks_ago=`date -v -14d -u +%F`
 
 for pod_index in {0..4}
 do
@@ -30,7 +38,6 @@ do
     ./sqlite-tools-osx/sqlite3 $pod/echo-sl-compare.sqlite "ALTER TABLE message RENAME TO message_echo;"
 
     chat_names=`kubectl -n echo-prod exec -ti $pod -- env | grep CHATS_NAMES_$pod_index | sed "s/CHATS_NAMES_$pod_index=//g" | sed 's/\[//g' | sed 's/\]//g' | sed 's/"/\x27/g'`
-    echo $chat_names
 
     ./sqlite-tools-osx/sqlite3 $pod/echo-sl-compare.sqlite "
 CREATE TABLE IF NOT EXISTS chat_sl (
@@ -89,9 +96,28 @@ CREATE TABLE IF NOT EXISTS
 (LIKE chat
 EXCLUDING ALL);
 
+CREATE TABLE IF NOT EXISTS
+    message_export 
+(LIKE message
+EXCLUDING ALL);
+
+CREATE INDEX IF NOT EXISTS idx_message_export_user_id ON message_export (user_id);
+
+CREATE TABLE IF NOT EXISTS
+    user_export 
+(LIKE \"user\"
+EXCLUDING ALL);
+
+CREATE UNIQUE INDEX IF NOT EXISTS idx_user_export_user_id ON user_export (user_id);
 
 DELETE FROM
     chat_export;
+
+DELETE FROM
+    message_export;
+
+DELETE FROM
+    user_export;
 
 INSERT INTO
     chat_export
@@ -101,14 +127,47 @@ FROM
     chat
 WHERE
     group_username in ($chat_names);
+
+INSERT INTO
+    message_export
+SELECT
+	a1.*
+FROM
+	message a1
+	INNER JOIN chat_export a2 ON a1.chat_id = a2.chat_id
+WHERE
+	a1.date > '$two_weeks_ago';
+
+INSERT INTO
+    user_export
+SELECT
+	a1.*
+FROM
+	\"user\" a1
+    INNER JOIN message_export a2 ON a1.user_id = a2.user_id
+ON CONFLICT DO NOTHING;
+
     " | ./pgsql/bin/psql -q $PG_URI
 
     ./pgsql/bin/pg_dump $PG_URI -t chat_export --column-inserts -a -f $pod/pgsql-dump-chat.sql
-    
-    sed -i.bak '/^INSERT/!d' $pod/pgsql-dump-chat.sql
-    sed -i.bak 's/public\.//g' $pod/pgsql-dump-chat.sql
-    sed -i.bak 's/chat_export/chat_sl/g' $pod/pgsql-dump-chat.sql
-
+    sed -i.0.bak '/^--/d' $pod/pgsql-dump-chat.sql
+    sed -i.1.bak '/^SET /d' $pod/pgsql-dump-chat.sql
+    sed -i.2.bak '/^SELECT pg_catalog/d' $pod/pgsql-dump-chat.sql
+    sed -i.3.bak 's/public\.chat_export/chat_sl/g' $pod/pgsql-dump-chat.sql
    ./sqlite-tools-osx/sqlite3 $pod/echo-sl-compare.sqlite ".read ${pod}/pgsql-dump-chat.sql"
+
+    ./pgsql/bin/pg_dump $PG_URI -t message_export --column-inserts -a -f $pod/pgsql-dump-message.sql
+    sed -i.0.bak '/^--/d' $pod/pgsql-dump-message.sql
+    sed -i.1.bak '/^SET /d' $pod/pgsql-dump-message.sql
+    sed -i.2.bak '/^SELECT pg_catalog/d' $pod/pgsql-dump-message.sql
+    sed -i.3.bak 's/public\.message_export/message_sl/g' $pod/pgsql-dump-message.sql
+   ./sqlite-tools-osx/sqlite3 $pod/echo-sl-compare.sqlite ".read ${pod}/pgsql-dump-message.sql"
+
+    ./pgsql/bin/pg_dump $PG_URI -t user_export --column-inserts -a -f $pod/pgsql-dump-user.sql
+    sed -i.0.bak '/^--/d' $pod/pgsql-dump-user.sql
+    sed -i.1.bak '/^SET /d' $pod/pgsql-dump-user.sql
+    sed -i.2.bak '/^SELECT pg_catalog/d' $pod/pgsql-dump-user.sql
+    sed -i.3.bak 's/public\.user_export/user_sl/g' $pod/pgsql-dump-user.sql
+   ./sqlite-tools-osx/sqlite3 $pod/echo-sl-compare.sqlite ".read ${pod}/pgsql-dump-user.sql"
  
 done
